requirements.txt
python-dotenv==1.0.1
openai==1.35.7
pydub==0.25.1
pysubs2==1.7.2